Update: Better project structure, renamed NarSimpleGUI to NarGUI.
