+++ src/main/java/org/opennars/applications/Launcher.java
@@ -120,7 +120,7 @@ public class Launcher extends javax.swing.JFrame {
     }// </editor-fold>//GEN-END:initComponents
 
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
-        org.opennars.applications.gui.NarSimpleGUI.main(null);
+        org.opennars.applications.nargui.NarGUI.main(null);
         this.dispose();
     }//GEN-LAST:event_jButton1ActionPerformed
