+++ src/main/java/org/opennars/applications/crossing/Camera.java
@@ -5,6 +5,8 @@
  */
 package org.opennars.applications.crossing;
 
+import org.opennars.applications.crossing.Encoders.InformPredictionNar;
+import org.opennars.applications.crossing.Entities.Entity;
 import java.util.List;
 import org.opennars.applications.Util;
 import org.opennars.main.Nar;
