+++ src/main/java/org/opennars/applications/crossing/Encoders/InformPredictionNar.java
@@ -21,11 +21,16 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.opennars.applications.crossing;
+package org.opennars.applications.crossing.Encoders;
 
+import org.opennars.applications.crossing.Entities.Entity;
+import org.opennars.applications.crossing.Entities.Car;
+import org.opennars.applications.crossing.Entities.Bike;
 import java.util.ArrayList;
 import java.util.List;
 import org.opennars.applications.Util;
+import org.opennars.applications.crossing.Crossing;
+import org.opennars.applications.crossing.TrafficLight;
 import org.opennars.main.Nar;
 
 public class InformPredictionNar {
