+++ src/main/java/org/opennars/applications/crossing/Entities/Bike.java
@@ -21,37 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.opennars.applications.crossing;
+package org.opennars.applications.crossing.Entities;
 
-import org.opennars.applications.Util;
-import org.opennars.entity.TruthValue;
-import processing.core.PApplet;
-
-public class Car extends Entity {
-
-    public Car(int id, double posX, double posY, double velocity, double angle, String label) {
-        this(id, posX, posY, velocity, angle);
-        this.label = label;
-    }
-    
-    public Car(int id, double posX, double posY, double velocity, double angle) {
-        super(id, posX, posY, velocity, angle);
-        maxSpeed = 2;
-    }
-
-    public void draw(PApplet applet, TruthValue truth, long time) {
-        float mul = Util.truthToValue(truth) * Util.timeToValue(time);
-        applet.fill(255, 0, 255, mul*255.0f);
-
-        if (!isPredicted && isAnomaly()) {
-            applet.stroke(255,0,0);
-        }
-        else {
-            applet.stroke(127);
-        }
-
-        super.draw(applet, truth, time);
-
-        applet.stroke(127);
+public class Bike extends Car {
+    public Bike(int id, double posX, double posY, double velocity, double angle, String label) {
+        super(id, posX, posY, velocity, angle, label);
     }
 }
