+++ src/main/java/org/opennars/applications/crossing/OperatorPanel.form
