+++ src/main/java/org/opennars/applications/nargui/NarGUI.form
